fixup! Can't navigate from disassembled source of a class file
